Added option to ignore URI parsing errors
When initializing a DataServiceContext there was a mandatory URI check
relying on .NETs Uri.IsWellFormedUriString. Sadly there is an active
issue on the .NET side that incorrectly flags URIs like
"http://192.168.0.1:1234/Instance/ODataV4/Company('123-Customer Place
Süd-Ost')/" as being invalid. The progress is tracked here:
https://github.com/dotnet/runtime/issues/21626
To give the user a simple option to prevent this issue from hindering
further development with the library I added a simple bool to ignore
this error.